Added the tab naviagation as well
