+++ src/app/_layout.tsx
@@ -1,7 +1,9 @@
-import { Stack } from "expo-router";
+import { Stack, Tabs } from "expo-router";
+import { Text } from "react-native";
+import { Feather } from "@expo/vector-icons";
 
 const Layout = () => {
     return <Stack />;
 };
 
-export default Layout;
+// export default Layout;
