+++ src/app/screens/(main)/HouseHold.tsx
@@ -1,0 +1,20 @@
+import { Text, View, StyleSheet } from "react-native";
+import { Stack } from "expo-router";
+
+const HouseHold = () => {
+    return (
+        <View style={styles.container}>
+            <Stack.Screen options={{ headerShown: false }} />
+            <Text>HouseHolding Screen</Text>
+        </View>
+    );
+};
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+    },
+});
+export default HouseHold;
