--- src/app/screens/RegisterScreen.tsx
@@ -1,111 +1,0 @@
-import React, { useState } from "react";
-import { StatusBar } from "expo-status-bar";
-import {
-  ActivityIndicator,
-  KeyboardAvoidingView,
-  StyleSheet,
-  Text,
-  TextInput,
-  View,
-} from "react-native";
-import { Link, Stack } from "expo-router";
-import { SafeAreaProvider } from "react-native-safe-area-context";
-import { Button, ThemeProvider } from "@rneui/themed";
-import { AppTheme, useStyles } from "../themes";
-import { firebaseAuth } from "../../../FirebaseConfig";
-import {
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-} from "firebase/auth";
-
-export function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [loading, setLoading] = useState(false);
-
-  const auth = firebaseAuth;
-
-  const signIn = async () => {
-    setLoading(true);
-    try {
-      const response = await signInWithEmailAndPassword(auth, email, password);
-      console.log(response);
-      // alert("Incorrect email or password")
-    } catch (error: any) {
-      console.log(error);
-      alert("Login failed: " + error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const signUp = async () => {
-    setLoading(true);
-    try {
-      const response = await createUserWithEmailAndPassword(
-        auth,
-        email,
-        password
-      );
-      console.log(response);
-      // alert("Incorrect email or password")
-    } catch (error: any) {
-      console.log(error);
-      alert("Registration failed: " + error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <View style={styles.container}>
-      <KeyboardAvoidingView behavior="padding">
-        <TextInput
-          value={email}
-          style={styles.input}
-          placeholder="Email"
-          autoCapitalize="none"
-          onChangeText={(text) => setEmail(text)}
-        ></TextInput>
-        <TextInput
-          value={password}
-          secureTextEntry={true}
-          style={styles.input}
-          placeholder="Password"
-          autoCapitalize="none"
-          onChangeText={(text) => setPassword(text)}
-        ></TextInput>
-        {loading ? (
-          <ActivityIndicator size="large" color="#0000ff" />
-        ) : (
-          <>
-            <Button title="Create Account" onPress={signUp}></Button>
-          </>
-        )}
-      </KeyboardAvoidingView>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  basicContainer: {
-    flex: 1,
-  },
-  buttonsContainer: {
-    width: 300,
-    paddingVertical: 100,
-    // position: "absolute",
-    // top: 70,
-  },
-  input: {
-    alignItems: "center",
-    justifyContent: "center",
-    padding: 10,
-  },
-});
